refactor(configs): refactor config files using @ConfigurationProperties
