+++ driver-service/src/main/java/com/cabaggregator/driverservice/DriverServiceApplication.java
@@ -2,8 +2,10 @@ package com.cabaggregator.driverservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class DriverServiceApplication {
 
     public static void main(String[] args) {
