+++ driver-service/src/main/java/com/cabaggregator/driverservice/config/MessageConfig.java
@@ -1,0 +1,15 @@
+package com.cabaggregator.driverservice.config;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Getter
+@AllArgsConstructor
+@ConfigurationProperties(prefix = "spring.messages.basename")
+public class MessageConfig {
+    private final String errorCauses;
+    private final String validationErrors;
+    private final String messages;
+    private final String encoding;
+}
