+++ driver-service/src/main/java/com/cabaggregator/driverservice/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.cabaggregator.driverservice.configuration;
+package com.cabaggregator.driverservice.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Contact;
