+++ driver-service/src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   messages:
-    encoding: "UTF-8"
     basename:
+      encoding: "UTF-8"
       error-causes: "classpath:error_causes"
       validation-errors: "classpath:validation_errors"
       messages: "classpath:messages"
