+++ passenger-service/src/main/java/com/cabaggregator/passengerservice/PassengerServiceApplication.java
@@ -2,8 +2,10 @@ package com.cabaggregator.passengerservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class PassengerServiceApplication {
 
     public static void main(String[] args) {
