+++ ride-service/src/main/java/com/cabaggregator/rideservice/RideServiceApplication.java
@@ -2,8 +2,10 @@ package com.cabaggregator.rideservice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class RideServiceApplication {
 
     public static void main(String[] args) {
